+++ CHANGELOG.md
@@ -1,3 +1,13 @@
+## 0.0.2
+
+* Added
+MotionContainer widget added
+
+* Changed 
+
+* Fixed
+Fix module tests
+
 ## 0.0.1
 
 * TODO: Describe initial release.